Flatten Terms
